refactoring worker scripts
